# Improve MathHelper performance