--- a/src/NCalc.Core/Helpers/MathHelper.cs
+++ b/src/NCalc.Core/Helpers/MathHelper.cs
@@ -413,10 +413,10 @@ public static class MathHelper
     {
         return value switch
         {
-            char when options is { DecimalAsDefault: true, AllowCharValues: false } => decimal.Parse(value.ToString()!, options.CultureInfo),
-            string when options is { DecimalAsDefault: true } => decimal.Parse(value.ToString()!, options.CultureInfo),
-            char when options is { AllowCharValues: false } => double.Parse(value.ToString()!, options.CultureInfo),
-            string => double.Parse(value.ToString()!, options.CultureInfo),
+            char ch when options is { DecimalAsDefault: true, AllowCharValues: false } => decimal.Parse(ch.ToString(), options.CultureInfo),
+            string s when options is { DecimalAsDefault: true } => decimal.Parse(s, options.CultureInfo),
+            char ch when options is { AllowCharValues: false } => double.Parse(ch.ToString(), options.CultureInfo),
+            string s => double.Parse(s, options.CultureInfo),
             bool boolean when options.AllowBooleanCalculation => boolean ? 1 : 0,
             _ => value
         };
@@ -427,7 +427,7 @@ public static class MathHelper
         return value switch
         {
             double @double => @double,
-            char => Convert.ToDouble(value.ToString(), options.CultureInfo),
+            char ch => Convert.ToDouble(ch.ToString(), options.CultureInfo),
             _ => Convert.ToDouble(value, options.CultureInfo)
         };
     }
@@ -437,7 +437,7 @@ public static class MathHelper
         return value switch
         {
             decimal @decimal => @decimal,
-            char => Convert.ToDecimal(value.ToString(), options.CultureInfo),
+            char ch => Convert.ToDecimal(ch.ToString(), options.CultureInfo),
             _ => Convert.ToDecimal(value, options.CultureInfo)
         };
     }
@@ -447,7 +447,7 @@ public static class MathHelper
         return value switch
         {
             int i => i,
-            char => Convert.ToInt32(value.ToString(), options.CultureInfo),
+            char ch => Convert.ToInt32(ch.ToString(), options.CultureInfo),
             _ => Convert.ToInt32(value, options.CultureInfo)
         };
     }