Improve MathHelper performance by avoiding boxing. 

BenchmarkDotNet=v0.13.12, OS=Windows 11 (10.0.22631.3527)
Intel Core i7-10700K (8C/16T), 1 CPU, 3.80GHz, 32GB RAM
.NET SDK=8.0.100
  [Host]     : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

| Method                      | Mean     | Error     | StdDev  | Ratio | Gen0   | Allocated |
|----------------------------|---------:|----------:|--------:|------:|-------:|----------:|
| ConvertIfNeededString_After | 24.80 ns | 0.110 ns | 0.103 ns | 1.00  | 0.0076 |      48 B |
| ConvertIfNeededString_Before | 34.60 ns | 0.150 ns | 0.140 ns | 1.40  | 0.0153 |      96 B |
| ConvertIfNeededChar_After | 29.90 ns | 0.120 ns | 0.112 ns | 1.00  | 0.0076 |      48 B | 
| ConvertIfNeededChar_Before  | 68.40 ns | 0.340 ns | 0.318 ns | 2.29  | 0.0305 |     192 B |